chore: extract timer and score/start to extra components
